refactor: extract getHotTracks()'s business logic to feature file, to separate it from db calls => so far, just approval tests pass
